refactor: Migrate Spotify demo data to generic demo data.
