+++ demos/spotify/src/main/java/com/guru/composecookbook/spotify/ui/details/components/BoxTopSection.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.guru.composecookbook.spotify.data.models.Album
+import com.guru.composecookbook.data.model.Album
 import com.guru.composecookbook.theme.typography
