+++ demos/spotify/src/main/java/com/guru/composecookbook/spotify/ui/details/components/SpotifyDetailScreen.kt
@@ -24,8 +24,9 @@ import androidx.compose.ui.res.imageResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import com.guru.composecookbook.data.AlbumsDataProvider
+import com.guru.composecookbook.data.model.Album
 import com.guru.composecookbook.spotify.data.SpotifyDataProvider
-import com.guru.composecookbook.spotify.data.models.Album
 import com.guru.composecookbook.theme.ComposeCookBookTheme
 import com.guru.composecookbook.theme.extensions.generateDominantColorState
 import com.guru.composecookbook.theme.modifiers.horizontalGradientBackground
@@ -99,7 +100,7 @@ fun BottomScrollableContent(scrollState: ScrollState, surfaceGradient: List<Colo
 @Preview
 @Composable
 fun PreviewDetailScreen() {
-    val album = SpotifyDataProvider.album
+    val album = AlbumsDataProvider.album
     ComposeCookBookTheme(true) {
         SpotifyDetailScreen(album = album)
     }
