+++ demos/spotify/src/main/java/com/guru/composecookbook/spotify/ui/home/components/SpotifyHome.kt
@@ -9,8 +9,6 @@ import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.FavoriteBorder
-import androidx.compose.material.icons.filled.PlayArrow
 import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -18,15 +16,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.guru.composecookbook.spotify.R
-import com.guru.composecookbook.theme.graySurface
 import com.guru.composecookbook.theme.typography
+import com.guru.composecookbook.data.AlbumsDataProvider
 import com.guru.composecookbook.spotify.data.SpotifyDataProvider
 import com.guru.composecookbook.verticalgrid.VerticalGrid
 import com.guru.composecookbook.theme.modifiers.horizontalGradientBackground
@@ -77,7 +71,7 @@ fun SpotifyTitle(text: String, modifier: Modifier = Modifier) {
 
 @Composable
 fun HomeGridSection() {
-    val items = remember { SpotifyDataProvider.albums }
+    val items = remember { AlbumsDataProvider.albums }
     VerticalGrid {
         items.take(6).forEach {
             SpotifyHomeGridItem(album = it)
@@ -87,7 +81,7 @@ fun HomeGridSection() {
 
 @Composable
 fun HomeLanesSection() {
-    val categories = remember { SpotifyDataProvider.listOfSpotifyHomeLanes }
+    val categories = remember { AlbumsDataProvider.listOfSpotifyHomeLanes }
     categories.forEachIndexed { index, lane ->
         SpotifyTitle(text = lane)
         SpotifyLane(index)
@@ -96,8 +90,8 @@ fun HomeLanesSection() {
 
 @Composable
 fun SpotifyLane(index: Int) {
-    val itemsEven = remember { SpotifyDataProvider.albums }
-    val itemsOdd = remember { SpotifyDataProvider.albums.asReversed() }
+    val itemsEven = remember { AlbumsDataProvider.albums }
+    val itemsOdd = remember { AlbumsDataProvider.albums.asReversed() }
     LazyRow {
         items(if (index % 2 == 0) itemsEven else itemsOdd) {
             SpotifyLaneItem(album = it)
